refactor(debug_server): extract codec endpoint family + send_json_resp shim (refs #332)
Third per-family extract from debug_server.c, stacked on the K144
(#336) and OTA (#337) extracts.

Endpoint owned by this PR:
  POST /codec/opus_test  — synthetic OPUS encoder smoke test (TT #264)

The opus_test endpoint was added in Wave 19 to bisect the SILK NSQ
stack-overflow crash; bisect found 24 KB as the watermark, leading to
the MIC_TASK_STACK_SIZE 16→32 KB bump in voice.c.  The endpoint stays
as a permanent diagnostic for codec-stack health checks + stack-budget
validation.

## Diff

  main/CMakeLists.txt           |   1 +
  main/debug_server.c           | 206 +-----  (codec block + 2 includes
                                                  + URI def + register call)
  main/debug_server_internal.h  |  11 ++  (new tab5_debug_send_json_resp)
  main/debug_server_codec.h     |  23 +++ (new)
  main/debug_server_codec.c     | 230 +++ (new)

debug_server.c:  4,202 → 4,023 LOC (−179).
Cumulative on this branch (K144 + OTA + codec): 4,520 → 4,023
= **−497 LOC** from origin/main baseline; **9 % shrinkage**.

## Infrastructure addition

This PR also exposes `tab5_debug_send_json_resp(req, root)` via
debug_server_internal.h.  ~57 in-file callers still use the static;
new family files (codec, future) go through the wrapper.  Cheap shim,
broadly useful for the remaining family extracts.

## Verification on hardware

build: clean — only debug_server_codec.c compiled.
flash: OK to /dev/ttyACM0, boot uptime 28 s.

POST /codec/opus_test?frames=10 → {
   encoder_active: true,
   frames_attempted: 10,
   frames_ok: 10,
   frames_failed: 0,
   total_bytes_out: 478,
   avg_bytes_per_frame: 47.8,
   us_per_frame: 2481.1,
}

OPUS encoder runs cleanly via the new module — same shape, same
counters as pre-extract.  Frames_ok=10/10, no failures, ~48 B/frame
matches the Wave 19 baseline.

story_smoke (14 steps): 14/14 PASS in 172 s.
