+++ main/debug_server.c
@@ -57,10 +57,11 @@
 #include "ui_wifi.h"
 #include "voice.h"
 /* Wave 23b (#332): per-family endpoint extracts + auth shim. */
+#include "debug_server_codec.h"
 #include "debug_server_internal.h"
 #include "debug_server_m5.h"
 #include "debug_server_ota.h"
-#include "widget.h"        /* Audit C4 (#202): widget_store_evictions_total */
+#include "widget.h" /* Audit C4 (#202): widget_store_evictions_total */
 #include "wifi.h"
 
 static const char *TAG = "debug_srv";
@@ -2812,6 +2813,11 @@ static esp_err_t send_json_resp(httpd_req_t *req, cJSON *root)
     return r;
 }
 
+/* Wave 23b (#332): public wrapper for the per-family extracts —
+ * see debug_server_internal.h.  The 57 in-file callers still use
+ * the static; new family files go through this. */
+esp_err_t tab5_debug_send_json_resp(httpd_req_t *req, struct cJSON *root) { return send_json_resp(req, (cJSON *)root); }
+
 /* ── GET /tasks — FreeRTOS task snapshot ────────────────────────── */
 /* Full per-task dump requires configUSE_TRACE_FACILITY + the
  * run-time-stats infra to be enabled in sdkconfig.  Most ESP-IDF
@@ -3531,195 +3537,7 @@ static esp_err_t debug_inject_audio_handler_impl(httpd_req_t *req) {
 
 esp_err_t debug_inject_audio_handler(httpd_req_t *req) { return debug_inject_audio_handler_impl(req); }
 
-/* TT #264 — synthetic OPUS encoder smoke test.  Drives
- * voice_codec_encode_uplink directly with N frames of synthetic
- * 16 kHz mono PCM (sine wave at 440 Hz, low amplitude).  Bypasses
- * the WS/mic pipeline so we can isolate "is the SILK NSQ encoder
- * working?" without coordinating Dragon's codec choice or a live mic.
- *
- * Runs the encode loop in a one-shot worker task with a 48 KB stack
- * (the issue reported 16 KB on voice_mic still triggers the SILK NSQ
- * panic — this is investigation step "expand stack until it works").
- *
- * Query params: ?frames=N (1-1000), ?heap=psram|internal|dma,
- * ?align=16|32|64, ?stack=N (KB, default 48).
- *
- * Returns counters: frames_attempted, frames_ok, frames_failed,
- * total_bytes_out.  If Tab5 panics mid-test the response never
- * lands; check uptime + reset_reason via /info to confirm. */
-#include "freertos/semphr.h"
-#include "voice_codec.h"
-
-typedef struct {
-   const int16_t *pcm;
-   size_t n_samples;
-   int frames;
-   int frames_ok;
-   int frames_failed;
-   uint32_t total_bytes_out;
-   int64_t dt_us;
-   SemaphoreHandle_t done;
-} opus_test_ctx_t;
-
-static void opus_test_task(void *arg) {
-   opus_test_ctx_t *ctx = (opus_test_ctx_t *)arg;
-   uint8_t out[1024];
-   int64_t t0 = esp_timer_get_time();
-   for (int i = 0; i < ctx->frames; i++) {
-      size_t out_len = 0;
-      esp_err_t e = voice_codec_encode_uplink(ctx->pcm, ctx->n_samples, out, sizeof(out), &out_len);
-      if (e == ESP_OK && out_len > 0) {
-         ctx->frames_ok++;
-         ctx->total_bytes_out += out_len;
-      } else {
-         ctx->frames_failed++;
-      }
-   }
-   ctx->dt_us = esp_timer_get_time() - t0;
-   xSemaphoreGive(ctx->done);
-   vTaskDelete(NULL);
-}
-static esp_err_t debug_opus_test_handler(httpd_req_t *req) {
-   if (!check_auth(req)) return ESP_OK;
-
-   /* Frame count: 50 frames = 1 second at 20 ms / frame.  Caller can
-    * override via ?frames=N (capped at 1000 to avoid DoS-ing a busy
-    * mic loop). */
-   int frames = 50;
-   {
-      char qbuf[64];
-      if (httpd_req_get_url_query_str(req, qbuf, sizeof(qbuf)) == ESP_OK) {
-         char val[16];
-         if (httpd_query_key_value(qbuf, "frames", val, sizeof(val)) == ESP_OK) {
-            int n = atoi(val);
-            if (n > 0 && n <= 1000) frames = n;
-         }
-      }
-   }
-
-   /* Allocate a 320-sample (20 ms @ 16 kHz) PCM frame.  The buffer
-    * heap_caps + alignment are part of what we're testing — issue
-    * #264 suspects the SILK encoder requires DMA-capable internal
-    * SRAM and/or 32/64-byte alignment.  Caller can override:
-    *   ?heap=psram  (default: same as the real mic buffer)
-    *   ?heap=internal  (try MALLOC_CAP_INTERNAL — investigation step 2)
-    *   ?align=N  (use heap_caps_aligned_alloc — step 1)
-    */
-   const size_t n_samples = 320;
-   int16_t *pcm = NULL;
-   uint32_t caps = MALLOC_CAP_SPIRAM | MALLOC_CAP_8BIT;
-   int align = 0;
-   {
-      char qbuf[64];
-      if (httpd_req_get_url_query_str(req, qbuf, sizeof(qbuf)) == ESP_OK) {
-         char val[16];
-         if (httpd_query_key_value(qbuf, "heap", val, sizeof(val)) == ESP_OK) {
-            if (strcmp(val, "internal") == 0) {
-               caps = MALLOC_CAP_INTERNAL | MALLOC_CAP_8BIT;
-            } else if (strcmp(val, "dma") == 0) {
-               caps = MALLOC_CAP_DMA | MALLOC_CAP_INTERNAL | MALLOC_CAP_8BIT;
-            }
-         }
-         if (httpd_query_key_value(qbuf, "align", val, sizeof(val)) == ESP_OK) {
-            int a = atoi(val);
-            if (a == 16 || a == 32 || a == 64) align = a;
-         }
-      }
-   }
-   if (align > 0) {
-      pcm = heap_caps_aligned_alloc((size_t)align, n_samples * sizeof(int16_t), caps);
-   } else {
-      pcm = heap_caps_malloc(n_samples * sizeof(int16_t), caps);
-   }
-   if (pcm == NULL) {
-      cJSON *r = cJSON_CreateObject();
-      cJSON_AddStringToObject(r, "error", "PSRAM alloc failed");
-      return send_json_resp(req, r);
-   }
-   /* Sine wave: 16 kHz sample rate, 440 Hz target.  Amplitude
-    * 0.1 × INT16_MAX = ~3276 (low to keep VBR honest). */
-   for (size_t i = 0; i < n_samples; i++) {
-      double t = (double)i / 16000.0;
-      double s = sin(2.0 * 3.14159265 * 440.0 * t);
-      pcm[i] = (int16_t)(s * 3276.0);
-   }
-
-   /* Force OPUS uplink — encoder spins up here on first call (will
-    * fall back to PCM silently if init fails). */
-   voice_codec_set_uplink(VOICE_CODEC_OPUS);
-   bool encoder_active = (voice_codec_get_uplink() == VOICE_CODEC_OPUS);
-
-   /* Pick the encode-task stack size from query (default 48 KB).
-    * Issue #264 reported voice_mic's 16 KB stack triggers SILK NSQ
-    * stack-protection-fault panic; this lets us bisect the minimum
-    * size that survives.  Cap at 96 KB to keep the alloc bounded. */
-   int stack_kb = 48;
-   {
-      char qbuf[64];
-      if (httpd_req_get_url_query_str(req, qbuf, sizeof(qbuf)) == ESP_OK) {
-         char val[16];
-         if (httpd_query_key_value(qbuf, "stack", val, sizeof(val)) == ESP_OK) {
-            int s = atoi(val);
-            if (s >= 8 && s <= 96) stack_kb = s;
-         }
-      }
-   }
-
-   /* Run the encode loop in a dedicated task with the requested
-    * stack — httpd's task stack is too small (4-8 KB depending on
-    * config) for SILK NSQ working storage.  Wait synchronously via
-    * semaphore. */
-   opus_test_ctx_t ctx = {
-       .pcm = pcm,
-       .n_samples = n_samples,
-       .frames = frames,
-       .done = xSemaphoreCreateBinary(),
-   };
-   if (ctx.done == NULL) {
-      heap_caps_free(pcm);
-      cJSON *r = cJSON_CreateObject();
-      cJSON_AddStringToObject(r, "error", "semaphore alloc failed");
-      return send_json_resp(req, r);
-   }
-   BaseType_t tc = xTaskCreate(opus_test_task, "opus_test", (uint32_t)stack_kb * 1024, &ctx, 5, NULL);
-   if (tc != pdPASS) {
-      vSemaphoreDelete(ctx.done);
-      heap_caps_free(pcm);
-      cJSON *r = cJSON_CreateObject();
-      cJSON_AddStringToObject(r, "error", "xTaskCreate failed");
-      cJSON_AddNumberToObject(r, "stack_kb", stack_kb);
-      return send_json_resp(req, r);
-   }
-   /* Wait up to 10 s for the task to finish.  At ~1 ms per frame
-    * even 1000 frames is well under this budget; if the task panics
-    * or hangs we'll fail-soft after 10 s with no payload to return. */
-   xSemaphoreTake(ctx.done, pdMS_TO_TICKS(10000));
-   vSemaphoreDelete(ctx.done);
-   int frames_ok = ctx.frames_ok;
-   int frames_failed = ctx.frames_failed;
-   uint32_t total_bytes_out = ctx.total_bytes_out;
-   int64_t dt_us = ctx.dt_us;
-   heap_caps_free(pcm);
-
-   cJSON *r = cJSON_CreateObject();
-   cJSON_AddBoolToObject(r, "encoder_active", encoder_active);
-   cJSON_AddStringToObject(r, "buf_heap", (caps & MALLOC_CAP_INTERNAL) ? "internal" : "psram");
-   cJSON_AddNumberToObject(r, "buf_align", align);
-   cJSON_AddNumberToObject(r, "task_stack_kb", stack_kb);
-   cJSON_AddNumberToObject(r, "frames_attempted", frames);
-   cJSON_AddNumberToObject(r, "frames_ok", frames_ok);
-   cJSON_AddNumberToObject(r, "frames_failed", frames_failed);
-   cJSON_AddNumberToObject(r, "total_bytes_out", (double)total_bytes_out);
-   cJSON_AddNumberToObject(r, "elapsed_ms", (double)(dt_us / 1000));
-   if (frames > 0) {
-      cJSON_AddNumberToObject(r, "avg_bytes_per_frame", (double)total_bytes_out / (double)frames);
-      cJSON_AddNumberToObject(r, "us_per_frame", (double)dt_us / (double)frames);
-   }
-   /* Restore PCM uplink so subsequent real WS traffic isn't
-    * unexpectedly OPUS-encoded. */
-   voice_codec_set_uplink(VOICE_CODEC_PCM);
-   return send_json_resp(req, r);
-}
+/* ── Codec endpoint family extracted to debug_server_codec.c (Wave 23b, #332) ─ */
 
 /* TT #328 Wave 12 — synthesize Wave 3 error.* events + their banner/
  * toast UX without needing real fault conditions.  POST body or query
@@ -4091,9 +3909,8 @@ esp_err_t tab5_debug_server_init(void)
     extern esp_err_t debug_inject_audio_handler(httpd_req_t * req);
     const httpd_uri_t uri_inject_audio = {
         .uri = "/debug/inject_audio", .method = HTTP_POST, .handler = debug_inject_audio_handler};
-    /* TT #264 — synthetic OPUS encoder smoke test. */
-    const httpd_uri_t uri_opus_test = {
-        .uri = "/codec/opus_test", .method = HTTP_POST, .handler = debug_opus_test_handler};
+    /* Wave 23b (#332): codec endpoint family — registered en-bloc via
+     * debug_server_codec_register() below. */
 
     /* TT #328 Wave 2 — synthetic WS-frame injector for error-surfacing
      * user-story tests (config_update.error, dictation_postprocessing_*,
@@ -4171,7 +3988,8 @@ esp_err_t tab5_debug_server_init(void)
     httpd_register_uri_handler(server, &uri_nvs_erase);
     httpd_register_uri_handler(server, &uri_inject_error);
     httpd_register_uri_handler(server, &uri_inject_audio);
-    httpd_register_uri_handler(server, &uri_opus_test);
+    /* Wave 23b (#332): codec endpoint family registered en-bloc. */
+    debug_server_codec_register(server);
     httpd_register_uri_handler(server, &uri_inject_ws);
 
     /* Log the URL */
